refactor: Share Spotify URL ID extraction logic
